backend/requirements.txt
gunicorn
flask
flask-cors
python-dotenv
requests